Replace tokio sockets by udp sockets
