+++ dds/Cargo.toml
@@ -26,7 +26,6 @@ network-interface = "1.1.1"
 fnmatch-regex = "0.2.0"
 
 tokio = { version = "1", features = [
-	"net",
 	"rt-multi-thread",
 	"sync",
 	"time",
